refactor(provider-arch): phase 2 sprint 6 - codex runtime extraction
Final structural extraction. Move the three big codex turn methods plus
the thread-id recovery helper from runtimeCore.ts into a dedicated
providers/codex/codexRuntime.ts module.

What moves to codexRuntime.ts (+1531 LOC):
- runCodexCli            (was: runCodexCliWithEvents — dispatcher between
                          exec/app-server modes with missing-thread retry +
                          fallback)
- runCodexAppServer      (was: runCodexAppServerWithEvents — WebSocket
                          JSON-RPC turn body, ~800 LOC)
- runCodexExecCli        (was: runCodexExecCliWithEvents — exec --json
                          stdout-streaming turn body, ~370 LOC)
- resolveCodexThreadId   (was: a private RuntimeCore method)
- CODEX_SANDBOX_MODE / CODEX_RUNTIME_MODE / CODEX_TURN_TIMEOUT_MS /
  CODEX_APP_SERVER_POST_TURN_QUIET_MS /
  CODEX_APP_SERVER_POST_TURN_DRAIN_TIMEOUT_MS env-driven constants
- AGENT_EXTRA_PATHS (codex-only copy)
- CodexRuntimeHost interface — narrow deps surface RuntimeCore exposes
  to codex turn code (runtimeEventLogger, permissionRouter,
  activeRunRegistry, coordinationStore, codexThreads Map, plus bound
  helpers: appendAgentMessage, listMessages, createPermission,
  decidePermission, resolveExecutionCwd, resolveSessionApprovalPolicy)

What changes in runtimeCore.ts (-1326 LOC, ~28% reduction):
- Three giant private methods + resolveCodexThreadId removed
- Five CODEX_* constants removed
- Imports added: runCodexCli, resolveCodexThreadId, CODEX_RUNTIME_MODE,
  type CodexRuntimeHost
- Thin codexHost getter constructs deps with bound methods on each call
- Call sites updated:
  - this.runCodexCliWithEvents(...) -> runCodexCli(this.codexHost, ...)
  - this.resolveCodexThreadId(...) -> resolveCodexThreadId(this.codexHost, ...)
- The codexThreads Map stays on RuntimeCore (cross-cutting cleanup
  via clearCodexThreadsForSession needs direct access)

After this PR runtimeCore.ts retains only orchestration call sites for
codex (dispatcher, thread-cache mutation, error-classification fallback).
All implementation code has moved out.

Validation:
- tsc --noEmit clean (aris-backend).
- vitest run --exclude e2e: 315/315 PASS (regression 0).
- Behaviour preserved end-to-end:
  - app-server mode initialize/thread/start/turn/start sequence intact
  - exec mode `codex exec --json` stdout streaming intact
  - missing-thread retry + fallback to exec mode unchanged
  - drain timeout + post-turn activity stabilisation unchanged
  - permission router binding/clearing unchanged
  - ndjson logging fields unchanged

Refs Phase 2 Sprint 6 of
docs/03-platform/codex-backend-alignment-plan.md.
