+++ services/aris-backend/src/runtime/runtimeCore.ts
@@ -49,6 +49,12 @@ import {
   normalizeCodexApprovalPolicy,
 } from './providers/codex/codexPermissionBridge.js';
 import {
+  CODEX_RUNTIME_MODE,
+  resolveCodexThreadId,
+  runCodexCli,
+  type CodexRuntimeHost,
+} from './providers/codex/codexRuntime.js';
+import {
   buildCodexPermissionKey,
   buildCodexThreadCacheKey,
   classifyCodexAppServerFailure,
@@ -119,8 +125,6 @@ const AGENT_EXTRA_PATHS = [
   '/home/ubuntu/.nvm/versions/node/v22.17.1/bin',
 ].join(':');
 const DEFAULT_APPROVAL_POLICY = normalizeApprovalPolicy(process.env.CODEX_APPROVAL_POLICY, 'on-request');
-const CODEX_SANDBOX_MODE = (process.env.CODEX_SANDBOX_MODE || 'workspace-write').trim();
-const CODEX_RUNTIME_MODE = (process.env.CODEX_RUNTIME_MODE || 'app-server').trim().toLowerCase();
 const HAPPY_MESSAGES_BATCH_LIMIT = 500;
 const HAPPY_MESSAGES_PAGE_MAX_LIMIT = 2000;
 const HAPPY_MESSAGES_MAX_PAGES = 1000;
@@ -141,27 +145,6 @@ const HAPPY_EVENT_LOG_MAX_BYTES = (() => {
 })();
 const HAPPY_MESSAGE_WRITE_BASE_DELAY_MS = 50;
 const HAPPY_MESSAGE_WRITE_MAX_RETRIES = 3;
-const CODEX_TURN_TIMEOUT_MS = (() => {
-  const parsed = Number.parseInt(process.env.CODEX_TURN_TIMEOUT_MS || '', 10);
-  if (Number.isFinite(parsed) && parsed >= 30_000) {
-    return parsed;
-  }
-  return 30 * 60 * 1000; // 30 minutes
-})();
-const CODEX_APP_SERVER_POST_TURN_QUIET_MS = (() => {
-  const parsed = Number.parseInt(process.env.CODEX_APP_SERVER_POST_TURN_QUIET_MS || '', 10);
-  if (Number.isFinite(parsed) && parsed >= 50) {
-    return parsed;
-  }
-  return 250;
-})();
-const CODEX_APP_SERVER_POST_TURN_DRAIN_TIMEOUT_MS = (() => {
-  const parsed = Number.parseInt(process.env.CODEX_APP_SERVER_POST_TURN_DRAIN_TIMEOUT_MS || '', 10);
-  if (Number.isFinite(parsed) && parsed >= 250) {
-    return parsed;
-  }
-  return 1_500;
-})();
 const STALE_RUN_TIMEOUT_MS = (() => {
   const parsed = Number.parseInt(process.env.HAPPY_STALE_RUN_TIMEOUT_MS || '', 10);
   if (Number.isFinite(parsed) && parsed >= 60_000) {
@@ -1196,6 +1179,22 @@ export class RuntimeCore {
   private readonly geminiSessionRegistry = new GeminiSessionRegistry();
   private readonly claudeSessionScanners = new Map<string, ClaudeSessionLogTracker>();
   private readonly codexThreads = new Map<string, string>();
+
+  private get codexHost(): CodexRuntimeHost {
+    return {
+      runtimeEventLogger: this.runtimeEventLogger,
+      permissionRouter: this.permissionRouter,
+      activeRunRegistry: this.activeRunRegistry,
+      coordinationStore: this.coordinationStore,
+      codexThreads: this.codexThreads,
+      appendAgentMessage: this.appendAgentMessage.bind(this),
+      listMessages: this.listMessages.bind(this),
+      createPermission: this.createPermission.bind(this),
+      decidePermission: this.decidePermission.bind(this),
+      resolveExecutionCwd: this.resolveExecutionCwd.bind(this),
+      resolveSessionApprovalPolicy: this.resolveSessionApprovalPolicy.bind(this),
+    };
+  }
   private readonly geminiPartialTextStates = new Map<string, GeminiPartialTextState>();
   private readonly coordinationStore: RuntimeCoordinationStore | null;
   private readonly permissionRouter: PermissionRouter;
@@ -2268,1281 +2267,6 @@ export class RuntimeCore {
     }
   }
 
-  private async runCodexCliWithEvents(
-    session: RuntimeSession,
-    prompt: string,
-    signal?: AbortSignal,
-    threadId?: string,
-    chatId?: string,
-    model?: string,
-    modelReasoningEffort?: ModelReasoningEffort,
-  ): Promise<{ output: string; cwd: string; streamedPersisted: boolean; agentMessagePersisted: boolean; threadId?: string }> {
-    if (CODEX_RUNTIME_MODE === 'exec') {
-      return this.runCodexExecCliWithEvents(session, prompt, signal, threadId, chatId, model, modelReasoningEffort);
-    }
-
-    try {
-      return await this.runCodexAppServerWithEvents(session, prompt, signal, threadId, chatId, model, modelReasoningEffort);
-    } catch (error) {
-      const failure = classifyCodexAppServerFailure(error);
-      if (failure.kind === 'missing_thread') {
-        throw error;
-      }
-      if (CODEX_RUNTIME_MODE === 'app-server-strict') {
-        throw error;
-      }
-
-      const threadCacheKey = buildCodexThreadCacheKey(session.id, chatId);
-      const hadThreadId = typeof threadId === 'string' && threadId.trim().length > 0;
-      if (failure.clearCachedThread) {
-        this.codexThreads.delete(threadCacheKey);
-      }
-      if (failure.retryWithFreshThread && hadThreadId) {
-        this.runtimeEventLogger.logParsed({
-          sessionId: session.id,
-          agent: 'codex',
-          ...(chatId ? { chatId } : {}),
-          model,
-          turnStatus: 'retrying',
-          channel: 'app_server',
-          stage: 'run_status',
-          payload: {
-            mode: 'app-server',
-            retryMode: 'fresh_thread',
-            failureKind: failure.kind,
-            errorMessage: failure.detail,
-          },
-        });
-        return this.runCodexAppServerWithEvents(session, prompt, signal, undefined, chatId, model, modelReasoningEffort);
-      }
-
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model,
-        turnStatus: 'fallback_to_exec',
-        channel: 'app_server',
-        stage: 'run_status',
-        payload: {
-          mode: 'app-server',
-          fallbackMode: 'exec',
-          failureKind: failure.kind,
-          errorMessage: failure.detail,
-          hadThreadId,
-        },
-      });
-      console.error(`codex app-server mode failed; falling back to exec mode [${failure.kind}]: ${failure.detail}`);
-      return this.runCodexExecCliWithEvents(session, prompt, signal, threadId, chatId, model, modelReasoningEffort);
-    }
-  }
-
-  private async runCodexAppServerWithEvents(
-    session: RuntimeSession,
-    prompt: string,
-    signal?: AbortSignal,
-    threadId?: string,
-    chatId?: string,
-    model?: string,
-    modelReasoningEffort?: ModelReasoningEffort,
-  ): Promise<{ output: string; cwd: string; streamedPersisted: boolean; agentMessagePersisted: boolean; threadId?: string }> {
-    const safeCwd = this.resolveExecutionCwd(session.metadata.path);
-    const threadCacheKey = buildCodexThreadCacheKey(session.id, chatId);
-    const sessionApprovalPolicy = this.resolveSessionApprovalPolicy(session);
-    const codexApprovalPolicy = normalizeCodexApprovalPolicy(sessionApprovalPolicy);
-    const selectedModel = normalizeModel(model) ?? resolveRuntimeModelSelection({
-      agent: 'codex',
-      sessionModel: session.metadata.model,
-    }).model;
-    const selectedReasoningEffort = normalizeModelReasoningEffort(modelReasoningEffort);
-    const autoApproveAll = sessionApprovalPolicy === 'yolo';
-    const effectiveSandboxMode = autoApproveAll ? 'danger-full-access' : CODEX_SANDBOX_MODE;
-    const mergedPath = `${process.env.PATH || ''}:${AGENT_EXTRA_PATHS}`;
-    const listenPort = await reserveCodexAppServerPort();
-    const listenUrl = buildCodexAppServerListenUrl(listenPort);
-    const args = [
-      ...(selectedModel ? ['-c', `model=${JSON.stringify(selectedModel)}`] : []),
-      ...(selectedReasoningEffort ? ['-c', `model_reasoning_effort=${JSON.stringify(selectedReasoningEffort)}`] : []),
-      'app-server',
-      '--listen',
-      listenUrl,
-    ];
-    const appServerPid = await launchDetachedCodexAppServerProcess({
-      cwd: safeCwd,
-      env: { ...process.env, PATH: mergedPath },
-      args,
-      signal,
-    });
-    this.runtimeEventLogger.logParsed({
-      sessionId: session.id,
-      agent: 'codex',
-      ...(chatId ? { chatId } : {}),
-      model: selectedModel,
-      turnStatus: 'run_started',
-      channel: 'app_server',
-      stage: 'run_status',
-      payload: {
-        mode: 'app-server',
-        args,
-        listenUrl,
-        appServerPid,
-      },
-    });
-
-    const socket = await connectCodexAppServerSocket(listenUrl, { signal });
-    let appendChain: Promise<void> = Promise.resolve();
-    let permissionChain: Promise<void> = Promise.resolve();
-    let lastAgentMessage = '';
-    let pendingAgentMessage = '';
-    let streamedPersisted = false;
-    let agentMessagePersisted = false;
-    let transportActivityTick = 0;
-    let lastTransportActivityAt = Date.now();
-    let resolvedThreadId = typeof threadId === 'string' && threadId.trim().length > 0
-      ? threadId.trim()
-      : '';
-    let activeTurnId = '';
-    let turnCompleted = false;
-    let runStatus: 'running' | 'completed' | 'failed' | 'aborted' | 'timed_out' = 'running';
-    let runErrorMessage: string | undefined;
-    let runFailureKind: CodexAppServerFailureKind | undefined;
-    const runtimePermissionIds = new Set<string>();
-    const persistedAgentMessageKeys = new Set<string>();
-
-    const pendingRequests = new Map<string, {
-      method: string;
-      resolve: (result: Record<string, unknown>) => void;
-      reject: (error: Error) => void;
-    }>();
-    let requestSequence = 0;
-
-    let resolveTurnCompletion: ((value: { status: string; errorMessage?: string }) => void) | null = null;
-    const turnCompletion = new Promise<{ status: string; errorMessage?: string }>((resolve) => {
-      resolveTurnCompletion = resolve;
-    });
-
-    const transportClosed = new Promise<never>((_, reject) => {
-      const handleClose = () => {
-        socket.removeEventListener('close', handleClose);
-        reject(new Error('codex app-server websocket closed before turn completion'));
-      };
-      socket.addEventListener('close', handleClose);
-    });
-
-    const abortController = createCodexAppServerAbortPromise({
-      signal,
-      pendingRequests,
-      onAbort: () => {
-        if (turnCompleted) {
-          return;
-        }
-        resolveTurnCompletion?.({ status: 'interrupted' });
-        try {
-          socket.close(1000, 'turn-abort');
-        } catch {
-          // ignore websocket close failures while aborting
-        }
-      },
-    });
-
-    const enqueueAppend = (
-      text: string,
-      meta: Record<string, unknown>,
-      options: { type?: string; title?: string } = {},
-    ) => {
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        channel: 'app_server',
-        stage: 'parsed_append',
-        payload: {
-          text,
-          meta,
-          options,
-        },
-      });
-      appendChain = appendChain
-        .then(() => this.appendAgentMessage(session.id, text, meta, options))
-        .catch((error) => {
-          const message = error instanceof Error ? error.message : String(error);
-          console.error(`failed to persist codex app-server event: ${message}`);
-        });
-    };
-
-    const sendJsonRpc = (payload: Record<string, unknown>): Promise<void> => new Promise((resolve, reject) => {
-      if (socket.readyState !== 1) {
-        reject(new Error('codex app-server websocket is not open'));
-        return;
-      }
-
-      try {
-        socket.send(JSON.stringify(payload));
-        resolve();
-      } catch (error) {
-        reject(error instanceof Error ? error : new Error(String(error)));
-      }
-    });
-
-    const sendJsonRpcResult = async (id: JsonRpcId | unknown, result: Record<string, unknown>) => {
-      await sendJsonRpc({
-        jsonrpc: '2.0',
-        id,
-        result,
-      });
-    };
-
-    const sendJsonRpcError = async (id: JsonRpcId | unknown, code: number, message: string) => {
-      await sendJsonRpc({
-        jsonrpc: '2.0',
-        id,
-        error: {
-          code,
-          message,
-        },
-      });
-    };
-
-    const sendRequest = <T extends Record<string, unknown> = Record<string, unknown>>(
-      method: string,
-      params: Record<string, unknown>,
-    ): Promise<T> => new Promise((resolve, reject) => {
-      const requestId = `aris-rpc-${requestSequence += 1}`;
-      const requestKey = toJsonRpcIdKey(requestId);
-      pendingRequests.set(requestKey, {
-        method,
-        resolve: (result) => resolve(result as T),
-        reject,
-      });
-
-      void sendJsonRpc({
-        jsonrpc: '2.0',
-        id: requestId,
-        method,
-        params,
-      }).catch((error) => {
-        pendingRequests.delete(requestKey);
-        reject(error instanceof Error ? error : new Error(String(error)));
-      });
-    });
-
-    const registerPermissionResponder = async (
-      key: string,
-      command: string,
-      reason: string,
-      risk: PermissionRisk,
-      responder: (decision: PermissionDecision) => Promise<void>,
-    ) => {
-      const knownPermissionId = this.permissionRouter.lookupCodexBinding(key);
-      if (knownPermissionId) {
-        const knownPermission = this.permissionRouter.getCachedPermission(knownPermissionId);
-        if (knownPermission?.state === 'pending') {
-          if (this.coordinationStore) {
-            this.permissionRouter.watchExternal({
-              permissionId: knownPermissionId,
-              responder,
-              signal,
-            });
-          } else {
-            this.permissionRouter.registerCodexResponder(knownPermissionId, responder);
-          }
-          runtimePermissionIds.add(knownPermissionId);
-          if (autoApproveAll) {
-            await this.decidePermission(knownPermissionId, 'allow_session');
-          }
-          return;
-        }
-        this.permissionRouter.clearCodexBinding(key);
-      }
-
-      const created = await this.createPermission({
-        sessionId: session.id,
-        ...(chatId ? { chatId } : {}),
-        agent: session.metadata.flavor === 'codex' ? 'codex' : 'unknown',
-        command,
-        reason,
-        risk,
-      });
-
-      this.permissionRouter.registerCodexBinding(key, created.id);
-      if (this.coordinationStore) {
-        this.permissionRouter.watchExternal({
-          permissionId: created.id,
-          responder,
-          signal,
-        });
-      } else {
-        this.permissionRouter.registerCodexResponder(created.id, responder);
-      }
-      runtimePermissionIds.add(created.id);
-      if (autoApproveAll) {
-        await this.decidePermission(created.id, 'allow_session');
-      }
-    };
-
-    const handleServerRequest = async (payload: Record<string, unknown>): Promise<void> => {
-      const method = asString(payload.method, '').trim();
-      const requestId = payload.id;
-      const params = asRecord(payload.params) ?? {};
-      const requestIdKey = toJsonRpcIdKey(requestId);
-
-      const approval = extractCodexAppServerApproval({
-        method,
-        params,
-        requestIdKey,
-        sessionId: session.id,
-      });
-      if (approval) {
-        await registerPermissionResponder(
-          buildScopedPermissionKey(approval.permissionKey, chatId),
-          approval.command,
-          approval.reason,
-          approval.risk,
-          (decision) => sendJsonRpcResult(requestId, { decision: approval.mapDecision(decision) }),
-        );
-        return;
-      }
-
-      if (method === 'mcpServer/elicitation/request') {
-        await sendJsonRpcResult(requestId, { action: 'cancel', content: null });
-        return;
-      }
-
-      if (method === 'item/tool/requestUserInput') {
-        await sendJsonRpcResult(requestId, { answers: {} });
-        return;
-      }
-
-      await sendJsonRpcError(requestId, -32601, `Unsupported server request method: ${method}`);
-    };
-
-    const handleServerNotification = (payload: Record<string, unknown>) => {
-      const method = asString(payload.method, '').trim();
-      const params = asRecord(payload.params) ?? {};
-
-      if (method === 'thread/started') {
-        const threadRecord = asRecord(params.thread);
-        const startedThreadId = asString(threadRecord?.id, '').trim();
-        if (startedThreadId) {
-          resolvedThreadId = startedThreadId;
-          this.codexThreads.set(threadCacheKey, startedThreadId);
-        }
-        return;
-      }
-
-      if (method === 'item/agentMessage/delta') {
-        const deltaRecord = asRecord(params.delta);
-        const deltaText = asString(
-          params.text,
-          asString(params.delta, asString(deltaRecord?.text, asString(deltaRecord?.delta, ''))),
-        );
-        if (deltaText) {
-          pendingAgentMessage += deltaText;
-          lastAgentMessage = pendingAgentMessage.trim();
-        }
-        return;
-      }
-
-      if (method === 'item/completed') {
-        const item = asRecord(params.item);
-        if (!item) {
-          return;
-        }
-
-        const itemType = asString(item.type, '');
-        if (itemType === 'agentMessage') {
-          const text = sanitizeAgentMessageText(asString(item.text, '').trim());
-          if (!text) {
-            return;
-          }
-          const itemTurnId = asString(params.turnId, activeTurnId).trim() || activeTurnId;
-          if (shouldSkipDuplicateAgentMessage(persistedAgentMessageKeys, itemTurnId, text)) {
-            return;
-          }
-          pendingAgentMessage = text;
-          lastAgentMessage = text;
-          streamedPersisted = true;
-          agentMessagePersisted = true;
-          enqueueAppend(
-            text,
-            {
-              ...(chatId ? { chatId } : {}),
-              requestedPath: session.metadata.path,
-              execCwd: safeCwd,
-              streamEvent: 'agent_message',
-              ...(resolvedThreadId ? { threadId: resolvedThreadId } : {}),
-            },
-            { type: 'message', title: 'Text Reply' },
-          );
-          return;
-        }
-
-        const fileWrite = inferCodexFileWriteItem(item);
-        if (fileWrite) {
-          const bodyParts = [`$ ${fileWrite.command || 'apply_patch'}`];
-          if (fileWrite.path) {
-            bodyParts.push(`path: ${fileWrite.path}`);
-          }
-          if (fileWrite.detail) {
-            bodyParts.push(fileWrite.detail);
-          }
-          if (shouldDisplayToolStatus(fileWrite.status)) {
-            bodyParts.push(`status: ${fileWrite.status}`);
-          }
-
-          streamedPersisted = true;
-          enqueueAppend(
-            bodyParts.join('\n'),
-            {
-              ...(chatId ? { chatId } : {}),
-              requestedPath: session.metadata.path,
-              execCwd: safeCwd,
-              actionType: 'file_write',
-              normalizedActionKind: 'file_write',
-              command: fileWrite.command,
-              path: fileWrite.path,
-              additions: fileWrite.additions,
-              deletions: fileWrite.deletions,
-              hasDiffSignal: fileWrite.hasDiffSignal,
-              streamEvent: 'file_change',
-              ...(resolvedThreadId ? { threadId: resolvedThreadId } : {}),
-            },
-            { type: 'tool', title: 'File Write' },
-          );
-          return;
-        }
-
-        if (itemType !== 'commandExecution') {
-          return;
-        }
-
-        const commandRaw = asString(item.command, '').trim();
-        const command = unwrapShellCommand(commandRaw);
-        const output = stripAnsi(asString(item.aggregatedOutput, '')).trim();
-        const exitCodeValue = item.exitCode;
-        const exitCode = typeof exitCodeValue === 'number' && Number.isFinite(exitCodeValue)
-          ? exitCodeValue
-          : null;
-        const actionType = inferActionTypeFromCommand(command);
-        const diffStats = actionType === 'file_write'
-          ? summarizeDiffText(output)
-          : { additions: 0, deletions: 0, hasDiffSignal: false };
-        const title = titleForActionType(actionType);
-        const bodyParts = [`$ ${command || 'command'}`];
-        if (output) {
-          bodyParts.push(output);
-        }
-        if (exitCode !== null) {
-          bodyParts.push(`exit code: ${exitCode}`);
-        }
-        const status = asString(item.status, '').trim();
-        if (shouldDisplayToolStatus(status)) {
-          bodyParts.push(`status: ${status}`);
-        }
-        const body = bodyParts.join('\n');
-
-        streamedPersisted = true;
-        enqueueAppend(
-          body,
-          {
-            ...(chatId ? { chatId } : {}),
-            requestedPath: session.metadata.path,
-            execCwd: safeCwd,
-            actionType,
-            normalizedActionKind: actionType,
-            command,
-            exitCode: exitCode ?? undefined,
-            additions: diffStats.additions,
-            deletions: diffStats.deletions,
-            hasDiffSignal: diffStats.hasDiffSignal,
-            streamEvent: 'command_execution',
-            ...(resolvedThreadId ? { threadId: resolvedThreadId } : {}),
-          },
-          {
-            type: 'tool',
-            title,
-          },
-        );
-        return;
-      }
-
-      if (method === 'turn/completed') {
-        if (!agentMessagePersisted) {
-          const recoveredText = sanitizeAgentMessageText(pendingAgentMessage.trim());
-          if (recoveredText) {
-            lastAgentMessage = recoveredText;
-            streamedPersisted = true;
-            agentMessagePersisted = true;
-            enqueueAppend(
-              recoveredText,
-              {
-                ...(chatId ? { chatId } : {}),
-                requestedPath: session.metadata.path,
-                execCwd: safeCwd,
-                streamEvent: 'agent_message_recovered',
-                ...(resolvedThreadId ? { threadId: resolvedThreadId } : {}),
-              },
-              { type: 'message', title: 'Text Reply' },
-            );
-          }
-        }
-
-        const turn = asRecord(params.turn);
-        const completedTurnId = asString(turn?.id, '').trim();
-        if (activeTurnId && completedTurnId && activeTurnId !== completedTurnId) {
-          return;
-        }
-
-        const status = asString(turn?.status, '').trim() || 'completed';
-        const errorMessage = asString(asRecord(turn?.error)?.message, '').trim() || undefined;
-        this.runtimeEventLogger.logParsed({
-          sessionId: session.id,
-          agent: 'codex',
-          ...(chatId ? { chatId } : {}),
-          model: selectedModel,
-          turnStatus: status,
-          channel: 'app_server',
-          stage: 'turn_status',
-          payload: {
-            turnId: completedTurnId || undefined,
-            threadId: resolvedThreadId || undefined,
-            ...(errorMessage ? { errorMessage } : {}),
-          },
-        });
-        turnCompleted = true;
-        resolveTurnCompletion?.({ status, errorMessage });
-      }
-    };
-
-    const handleIncomingPayloadText = (rawText: string) => {
-      transportActivityTick += 1;
-      lastTransportActivityAt = Date.now();
-      const rawLine = rawText.trim();
-      this.runtimeEventLogger.logRaw({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        channel: 'app_server',
-        line: rawLine,
-      });
-      const payload = parseJsonLine(rawLine);
-      if (!payload) {
-        this.runtimeEventLogger.logParsed({
-          sessionId: session.id,
-          agent: 'codex',
-          ...(chatId ? { chatId } : {}),
-          model: selectedModel,
-          channel: 'app_server',
-          stage: 'incoming_payload',
-          payload: { parseError: 'invalid_json' },
-        });
-        return;
-      }
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        channel: 'app_server',
-        stage: 'incoming_payload',
-        payload,
-      });
-
-      const messageMethod = typeof payload.method === 'string' ? payload.method : '';
-      const hasId = Object.prototype.hasOwnProperty.call(payload, 'id');
-
-      if (messageMethod && hasId) {
-        permissionChain = permissionChain
-          .then(() => handleServerRequest(payload))
-          .catch((error) => {
-            const message = error instanceof Error ? error.message : String(error);
-            console.error(`failed to handle codex app-server request: ${message}`);
-          });
-        return;
-      }
-
-      if (messageMethod) {
-        handleServerNotification(payload);
-        return;
-      }
-
-      if (!hasId) {
-        return;
-      }
-
-      const idKey = toJsonRpcIdKey(payload.id);
-      const pending = pendingRequests.get(idKey);
-      if (!pending) {
-        return;
-      }
-      pendingRequests.delete(idKey);
-
-      const errorPayload = asRecord(payload.error);
-      if (errorPayload) {
-        const rpcMessage = asString(errorPayload.message, `JSON-RPC ${pending.method} failed`);
-        pending.reject(new Error(rpcMessage));
-        return;
-      }
-
-      const resultPayload = asRecord(payload.result) ?? {};
-      pending.resolve(resultPayload);
-    };
-
-    socket.addEventListener('message', (event) => {
-      handleIncomingPayloadText(normalizeCodexAppServerMessageData(event.data));
-    });
-    socket.addEventListener('close', () => {
-      rejectCodexAppServerPendingRequests(pendingRequests, 'codex app-server websocket closed before turn completion');
-    });
-
-    const closeChild = async () => {
-      try {
-        socket.close(1000, 'turn-complete');
-      } catch {
-        // ignore websocket close failures during shutdown
-      }
-      const pseudoChild = { pid: appServerPid, killed: false, kill: () => false };
-      terminateCodexAppServerProcess(pseudoChild, process.kill, 'SIGTERM');
-      await new Promise((resolve) => setTimeout(resolve, 250));
-      terminateCodexAppServerProcess(pseudoChild, process.kill, 'SIGKILL');
-    };
-
-    const waitForPostTurnDrain = async () => {
-      const drainDeadline = Date.now() + CODEX_APP_SERVER_POST_TURN_DRAIN_TIMEOUT_MS;
-
-      while (Date.now() < drainDeadline) {
-        await waitForStableActivity({
-          getActivityTick: () => transportActivityTick,
-          getLastActivityAt: () => lastTransportActivityAt,
-          quietMs: CODEX_APP_SERVER_POST_TURN_QUIET_MS,
-          timeoutMs: Math.max(10, drainDeadline - Date.now()),
-        });
-
-        const activityTickSnapshot = transportActivityTick;
-        const appendSnapshot = appendChain;
-        const permissionSnapshot = permissionChain;
-        await appendSnapshot;
-        await permissionSnapshot;
-
-        if (
-          activityTickSnapshot === transportActivityTick
-          && appendSnapshot === appendChain
-          && permissionSnapshot === permissionChain
-          && Date.now() - lastTransportActivityAt >= CODEX_APP_SERVER_POST_TURN_QUIET_MS
-        ) {
-          return;
-        }
-      }
-
-      await appendChain;
-      await permissionChain;
-    };
-
-    try {
-      await sendRequest('initialize', {
-        clientInfo: {
-          name: 'aris-runtime',
-          title: 'ARIS Runtime',
-          version: '0.1.0',
-        },
-        capabilities: {
-          experimentalApi: true,
-          optOutNotificationMethods: [
-            'item/commandExecution/outputDelta',
-            'item/commandExecution/terminalInteraction',
-          ],
-        },
-      });
-      await sendJsonRpc({
-        jsonrpc: '2.0',
-        method: 'initialized',
-        params: {},
-      });
-
-      if (resolvedThreadId) {
-        const resumed = await sendRequest('thread/resume', {
-          threadId: resolvedThreadId,
-          cwd: safeCwd,
-          approvalPolicy: codexApprovalPolicy,
-          sandbox: effectiveSandboxMode,
-          persistExtendedHistory: true,
-        });
-        const resumedThreadId = asString(asRecord(resumed.thread)?.id, '').trim();
-        if (resumedThreadId) {
-          resolvedThreadId = resumedThreadId;
-          this.codexThreads.set(threadCacheKey, resumedThreadId);
-        }
-      } else {
-        const started = await sendRequest('thread/start', {
-          cwd: safeCwd,
-          approvalPolicy: codexApprovalPolicy,
-          sandbox: effectiveSandboxMode,
-          experimentalRawEvents: false,
-          persistExtendedHistory: true,
-        });
-        const startedThreadId = asString(asRecord(started.thread)?.id, '').trim();
-        if (startedThreadId) {
-          resolvedThreadId = startedThreadId;
-          this.codexThreads.set(threadCacheKey, startedThreadId);
-        }
-      }
-
-      if (!resolvedThreadId) {
-        throw new Error('codex app-server did not return a thread id');
-      }
-
-      const turnStarted = await sendRequest('turn/start', {
-        threadId: resolvedThreadId,
-        input: [
-          {
-            type: 'text',
-            text: prompt,
-            text_elements: [],
-          },
-        ],
-        approvalPolicy: codexApprovalPolicy,
-      });
-      activeTurnId = asString(asRecord(turnStarted.turn)?.id, '').trim();
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        turnStatus: 'turn_started',
-        channel: 'app_server',
-        stage: 'turn_status',
-        payload: {
-          turnId: activeTurnId || undefined,
-          threadId: resolvedThreadId || undefined,
-          timeoutMs: CODEX_TURN_TIMEOUT_MS,
-        },
-      });
-
-      let turnTimeout: NodeJS.Timeout | undefined;
-      const completion = await Promise.race([
-        turnCompletion,
-        ...(abortController.interrupted ? [abortController.interrupted] : []),
-        transportClosed,
-        new Promise<{ status: string; errorMessage?: string }>((_resolve, reject) => {
-          turnTimeout = setTimeout(() => {
-            runStatus = 'timed_out';
-            runErrorMessage = `turn timeout exceeded (${CODEX_TURN_TIMEOUT_MS}ms)`;
-            reject(new Error(`codex app-server turn timed out after ${CODEX_TURN_TIMEOUT_MS}ms`));
-          }, CODEX_TURN_TIMEOUT_MS);
-        }),
-      ]).finally(() => {
-        if (turnTimeout) {
-          clearTimeout(turnTimeout);
-        }
-      });
-
-      await appendChain;
-      await permissionChain;
-
-      const finalText = sanitizeAgentMessageText(lastAgentMessage.trim());
-      if (signal?.aborted || completion.status === 'interrupted') {
-        runStatus = 'aborted';
-        return {
-          output: trimOutput(finalText),
-          cwd: safeCwd,
-          streamedPersisted,
-          agentMessagePersisted,
-          threadId: resolvedThreadId || undefined,
-        };
-      }
-
-      if (completion.status === 'failed' && !finalText) {
-        const suffix = completion.errorMessage ? `: ${completion.errorMessage}` : '';
-        runStatus = 'failed';
-        runErrorMessage = completion.errorMessage;
-        throw new Error(`codex app-server turn failed${suffix}`);
-      }
-
-      runStatus = 'completed';
-      return {
-        output: trimOutput(finalText),
-        cwd: safeCwd,
-        streamedPersisted,
-        agentMessagePersisted,
-        threadId: resolvedThreadId || undefined,
-      };
-    } catch (error) {
-      const failure = classifyCodexAppServerFailure(error);
-      if (runStatus === 'running') {
-        runStatus = signal?.aborted ? 'aborted' : 'failed';
-        runErrorMessage = failure.detail;
-      }
-      runFailureKind = failure.kind;
-      throw error;
-    } finally {
-      abortController.dispose();
-      rejectCodexAppServerPendingRequests(pendingRequests, 'The operation was aborted');
-
-      await permissionChain.catch(() => undefined);
-      await appendChain.catch(() => undefined);
-
-      await this.permissionRouter.finalizeCodexPermissions(runtimePermissionIds, {
-        preservePending: this.activeRunRegistry.isDraining() && !signal?.aborted,
-      });
-
-      if (!turnCompleted && !signal?.aborted) {
-        if (runStatus === 'running') {
-          runStatus = 'failed';
-          runErrorMessage = 'turn did not complete before process close';
-          runFailureKind = 'websocket_closed';
-        }
-        this.runtimeEventLogger.logParsed({
-          sessionId: session.id,
-          agent: 'codex',
-          ...(chatId ? { chatId } : {}),
-          model: selectedModel,
-          turnStatus: 'turn_incomplete',
-          channel: 'app_server',
-          stage: 'turn_status',
-          payload: {
-            threadId: resolvedThreadId || undefined,
-            turnId: activeTurnId || undefined,
-            listenUrl,
-            appServerPid,
-            ...(runFailureKind ? { failureKind: runFailureKind } : {}),
-            ...(runErrorMessage ? { errorMessage: runErrorMessage } : {}),
-          },
-        });
-        if (runFailureKind === 'websocket_closed' || (!runFailureKind && runStatus === 'failed')) {
-          console.error(`codex app-server transport closed early; pid=${appServerPid} listenUrl=${listenUrl}`);
-        }
-      }
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        turnStatus: runStatus,
-        channel: 'app_server',
-        stage: 'run_status',
-        payload: {
-          threadId: resolvedThreadId || undefined,
-          turnId: activeTurnId || undefined,
-          turnCompleted,
-          ...(runFailureKind ? { failureKind: runFailureKind } : {}),
-          ...(runErrorMessage ? { errorMessage: runErrorMessage } : {}),
-        },
-      });
-
-      await closeChild();
-    }
-  }
-
-  private async runCodexExecCliWithEvents(
-    session: RuntimeSession,
-    prompt: string,
-    signal?: AbortSignal,
-    threadId?: string,
-    chatId?: string,
-    model?: string,
-    modelReasoningEffort?: ModelReasoningEffort,
-  ): Promise<{ output: string; cwd: string; streamedPersisted: boolean; agentMessagePersisted: boolean; threadId?: string }> {
-    const safeCwd = this.resolveExecutionCwd(session.metadata.path);
-    const threadCacheKey = buildCodexThreadCacheKey(session.id, chatId);
-    const sessionApprovalPolicy = this.resolveSessionApprovalPolicy(session);
-    const codexApprovalPolicy = normalizeCodexApprovalPolicy(sessionApprovalPolicy);
-    const selectedModel = normalizeModel(model) ?? resolveRuntimeModelSelection({
-      agent: 'codex',
-      sessionModel: session.metadata.model,
-    }).model;
-    const selectedReasoningEffort = normalizeModelReasoningEffort(modelReasoningEffort);
-    const autoApproveAll = sessionApprovalPolicy === 'yolo';
-    const effectiveSandboxMode = autoApproveAll ? 'danger-full-access' : CODEX_SANDBOX_MODE;
-    const mergedPath = `${process.env.PATH || ''}:${AGENT_EXTRA_PATHS}`;
-    const execArgs = threadId
-      ? ['exec', 'resume', threadId, '--json', prompt]
-      : ['exec', '--json', prompt];
-    const args = [
-      '-a',
-      codexApprovalPolicy,
-      '-s',
-      effectiveSandboxMode,
-      ...(selectedModel ? ['-m', selectedModel] : []),
-      ...(selectedReasoningEffort ? ['-c', `model_reasoning_effort=${JSON.stringify(selectedReasoningEffort)}`] : []),
-      ...execArgs,
-    ];
-    const child = spawn('codex', args, {
-      cwd: safeCwd,
-      env: { ...process.env, PATH: mergedPath },
-      stdio: ['pipe', 'pipe', 'pipe'],
-      signal,
-    });
-    this.runtimeEventLogger.logParsed({
-      sessionId: session.id,
-      agent: 'codex',
-      ...(chatId ? { chatId } : {}),
-      model: selectedModel,
-      turnStatus: 'run_started',
-      channel: 'exec_cli',
-      stage: 'run_status',
-      payload: {
-        mode: 'exec',
-        args,
-      },
-    });
-
-    const stdoutLines = createInterface({ input: child.stdout });
-    let stderr = '';
-    let appendChain: Promise<void> = Promise.resolve();
-    let permissionChain: Promise<void> = Promise.resolve();
-    let lastAgentMessage = '';
-    let streamedPersisted = false;
-    let agentMessagePersisted = false;
-    let resolvedThreadId = typeof threadId === 'string' && threadId.trim().length > 0
-      ? threadId.trim()
-      : '';
-    const persistedAgentMessageKeys = new Set<string>();
-
-    const enqueueAppend = (
-      text: string,
-      meta: Record<string, unknown>,
-      options: { type?: string; title?: string } = {},
-    ) => {
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        channel: 'exec_cli',
-        stage: 'parsed_append',
-        payload: {
-          text,
-          meta,
-          options,
-        },
-      });
-      appendChain = appendChain
-        .then(() => this.appendAgentMessage(session.id, text, meta, options))
-        .catch((error) => {
-          const message = error instanceof Error ? error.message : String(error);
-          console.error(`failed to persist codex stream event: ${message}`);
-        });
-    };
-
-    child.stderr.on('data', (chunk: Buffer | string) => {
-      stderr += typeof chunk === 'string' ? chunk : chunk.toString('utf8');
-    });
-
-    const enqueuePermission = (request: CodexPermissionRequest) => {
-      permissionChain = permissionChain
-        .then(async () => {
-          const key = buildScopedPermissionKey(buildCodexPermissionKey(session.id, request), chatId);
-          const knownPermissionId = this.permissionRouter.lookupCodexBinding(key);
-          if (knownPermissionId) {
-            const knownPermission = this.permissionRouter.getCachedPermission(knownPermissionId);
-            if (knownPermission?.state === 'pending') {
-              if (autoApproveAll) {
-                await this.decidePermission(knownPermissionId, 'allow_session');
-              }
-              return;
-            }
-            this.permissionRouter.clearCodexBinding(key);
-          }
-
-          const created = await this.createPermission({
-            sessionId: session.id,
-            ...(chatId ? { chatId } : {}),
-            agent: session.metadata.flavor === 'codex' ? 'codex' : 'unknown',
-            command: request.command,
-            reason: request.reason,
-            risk: request.risk,
-          });
-
-          this.permissionRouter.registerCodexBinding(key, created.id);
-          if (autoApproveAll) {
-            await this.decidePermission(created.id, 'allow_session');
-          }
-        })
-        .catch((error) => {
-          const message = error instanceof Error ? error.message : String(error);
-          console.error(`failed to create codex permission request: ${message}`);
-        });
-    };
-
-    stdoutLines.on('line', (line) => {
-      const rawLine = line.trim();
-      this.runtimeEventLogger.logRaw({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        channel: 'exec_cli',
-        line: rawLine,
-      });
-      const payload = parseJsonLine(rawLine);
-      if (!payload) {
-        this.runtimeEventLogger.logParsed({
-          sessionId: session.id,
-          agent: 'codex',
-          ...(chatId ? { chatId } : {}),
-          model: selectedModel,
-          channel: 'exec_cli',
-          stage: 'incoming_payload',
-          payload: { parseError: 'invalid_json' },
-        });
-        return;
-      }
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        channel: 'exec_cli',
-        stage: 'incoming_payload',
-        payload,
-      });
-
-      if (payload.type === 'thread.started') {
-        const startedThreadId = asString(payload.thread_id, '').trim();
-        if (startedThreadId) {
-          resolvedThreadId = startedThreadId;
-          this.codexThreads.set(threadCacheKey, startedThreadId);
-        }
-        return;
-      }
-
-      const approvalRequest = extractCodexPermissionRequest(payload);
-      if (approvalRequest) {
-        enqueuePermission(approvalRequest);
-        return;
-      }
-
-      if (payload.type !== 'item.completed') {
-        return;
-      }
-
-      const item = asRecord(payload.item);
-      if (!item) {
-        return;
-      }
-
-      const itemType = asString(item.type, '');
-      if (itemType === 'agent_message') {
-        const text = sanitizeAgentMessageText(asString(item.text, '').trim());
-        if (text) {
-          const itemTurnId = asString(
-            payload.turn_id,
-            asString(payload.turnId, asString(item.turn_id, asString(item.turnId, ''))),
-          ).trim();
-          if (shouldSkipDuplicateAgentMessage(persistedAgentMessageKeys, itemTurnId, text)) {
-            return;
-          }
-          lastAgentMessage = text;
-          streamedPersisted = true;
-          agentMessagePersisted = true;
-          enqueueAppend(
-            text,
-            {
-              ...(chatId ? { chatId } : {}),
-              requestedPath: session.metadata.path,
-              execCwd: safeCwd,
-              streamEvent: 'agent_message',
-              ...(resolvedThreadId ? { threadId: resolvedThreadId } : {}),
-            },
-            { type: 'message', title: 'Text Reply' },
-          );
-        }
-        return;
-      }
-
-      const fileWrite = inferCodexFileWriteItem(item);
-      if (fileWrite) {
-        const bodyParts = [`$ ${fileWrite.command || 'apply_patch'}`];
-        if (fileWrite.path) {
-          bodyParts.push(`path: ${fileWrite.path}`);
-        }
-        if (fileWrite.detail) {
-          bodyParts.push(fileWrite.detail);
-        }
-        if (shouldDisplayToolStatus(fileWrite.status)) {
-          bodyParts.push(`status: ${fileWrite.status}`);
-        }
-
-        streamedPersisted = true;
-        enqueueAppend(
-          bodyParts.join('\n'),
-          {
-            ...(chatId ? { chatId } : {}),
-            requestedPath: session.metadata.path,
-            execCwd: safeCwd,
-            actionType: 'file_write',
-            normalizedActionKind: 'file_write',
-            command: fileWrite.command,
-            path: fileWrite.path,
-            additions: fileWrite.additions,
-            deletions: fileWrite.deletions,
-            hasDiffSignal: fileWrite.hasDiffSignal,
-            streamEvent: 'file_change',
-            ...(resolvedThreadId ? { threadId: resolvedThreadId } : {}),
-          },
-          { type: 'tool', title: 'File Write' },
-        );
-        return;
-      }
-
-      if (itemType !== 'command_execution') {
-        return;
-      }
-
-      const commandRaw = asString(item.command, '').trim();
-      const command = unwrapShellCommand(commandRaw);
-      const output = stripAnsi(asString(item.aggregated_output, '')).trim();
-      const exitCodeValue = item.exit_code;
-      const exitCode = typeof exitCodeValue === 'number' && Number.isFinite(exitCodeValue)
-        ? exitCodeValue
-        : null;
-      const actionType = inferActionTypeFromCommand(command);
-      const diffStats = actionType === 'file_write'
-        ? summarizeDiffText(output)
-        : { additions: 0, deletions: 0, hasDiffSignal: false };
-      const title = titleForActionType(actionType);
-      const bodyParts = [`$ ${command || 'command'}`];
-      if (output) {
-        bodyParts.push(output);
-      }
-      if (exitCode !== null) {
-        bodyParts.push(`exit code: ${exitCode}`);
-      }
-      const body = bodyParts.join('\n');
-
-      streamedPersisted = true;
-      enqueueAppend(
-        body,
-        {
-          ...(chatId ? { chatId } : {}),
-          requestedPath: session.metadata.path,
-          execCwd: safeCwd,
-          actionType,
-          normalizedActionKind: actionType,
-          command,
-          exitCode: exitCode ?? undefined,
-          additions: diffStats.additions,
-          deletions: diffStats.deletions,
-          hasDiffSignal: diffStats.hasDiffSignal,
-          streamEvent: 'command_execution',
-          ...(resolvedThreadId ? { threadId: resolvedThreadId } : {}),
-        },
-        {
-          type: 'tool',
-          title,
-        },
-      );
-    });
-
-    const result = await new Promise<{ code: number | null; signal: NodeJS.Signals | null }>((resolve, reject) => {
-      child.once('error', reject);
-      child.once('close', (code, signal) => resolve({ code, signal }));
-    });
-
-    await appendChain;
-    await permissionChain;
-
-    const finalText = sanitizeAgentMessageText(lastAgentMessage.trim());
-    if (signal?.aborted) {
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        turnStatus: 'aborted',
-        channel: 'exec_cli',
-        stage: 'run_status',
-        payload: {
-          threadId: resolvedThreadId || undefined,
-        },
-      });
-      return {
-        output: trimOutput(finalText),
-        cwd: safeCwd,
-        streamedPersisted,
-        agentMessagePersisted,
-        threadId: resolvedThreadId || undefined,
-      };
-    }
-
-    if (result.code !== 0 && !finalText) {
-      const detail = stripAnsi(stderr).slice(0, 800) || `exit code ${result.code}`;
-      this.runtimeEventLogger.logParsed({
-        sessionId: session.id,
-        agent: 'codex',
-        ...(chatId ? { chatId } : {}),
-        model: selectedModel,
-        turnStatus: 'failed',
-        channel: 'exec_cli',
-        stage: 'run_status',
-        payload: {
-          threadId: resolvedThreadId || undefined,
-          exitCode: result.code,
-          errorMessage: detail,
-        },
-      });
-      throw new Error(`codex CLI failed: ${detail}`);
-    }
-
-    this.runtimeEventLogger.logParsed({
-      sessionId: session.id,
-      agent: 'codex',
-      ...(chatId ? { chatId } : {}),
-      model: selectedModel,
-      turnStatus: 'completed',
-      channel: 'exec_cli',
-      stage: 'run_status',
-      payload: {
-        threadId: resolvedThreadId || undefined,
-        exitCode: result.code,
-      },
-    });
-    return {
-      output: trimOutput(finalText),
-      cwd: safeCwd,
-      streamedPersisted,
-      agentMessagePersisted,
-      threadId: resolvedThreadId || undefined,
-    };
-  }
-
-  private async resolveCodexThreadId(sessionId: string, chatId?: string): Promise<string | undefined> {
-    const cacheKey = buildCodexThreadCacheKey(sessionId, chatId);
-    const cached = this.codexThreads.get(cacheKey);
-    if (cached) {
-      return cached;
-    }
-
-    try {
-      const history = await this.listMessages(sessionId);
-      for (let index = history.length - 1; index >= 0; index -= 1) {
-        if (chatId) {
-          const rawChatId = history[index]?.meta?.chatId;
-          const messageChatId = typeof rawChatId === 'string' ? rawChatId.trim() : '';
-          if (messageChatId !== chatId) {
-            continue;
-          }
-        }
-        const candidate = history[index]?.meta?.threadId;
-        if (typeof candidate !== 'string') {
-          continue;
-        }
-        const trimmed = candidate.trim();
-        if (!trimmed) {
-          continue;
-        }
-        this.codexThreads.set(cacheKey, trimmed);
-        return trimmed;
-      }
-    } catch {
-      // Ignore thread recovery failures and start a new Codex thread.
-    }
-
-    return undefined;
-  }
-
   private async resolveClaudeThreadId(sessionId: string, chatId?: string): Promise<string | undefined> {
     try {
       return recoverClaudeThreadIdFromMessages(await this.listMessages(sessionId), chatId);
@@ -3784,9 +2508,9 @@ export class RuntimeCore {
       };
 
       if (isCodex) {
-        const recoveredThreadId = preferredThreadId ?? await this.resolveCodexThreadId(session.id, scopedChatId);
+        const recoveredThreadId = preferredThreadId ?? await resolveCodexThreadId(this.codexHost, session.id, scopedChatId);
         try {
-          response = await this.runCodexCliWithEvents(
+          response = await runCodexCli(this.codexHost, 
             session,
             prompt,
             controller.signal,
@@ -3802,7 +2526,7 @@ export class RuntimeCore {
 
           // Stored thread id became invalid; clear and start a fresh Codex thread.
           this.codexThreads.delete(threadCacheKey);
-          response = await this.runCodexCliWithEvents(
+          response = await runCodexCli(this.codexHost, 
             session,
             prompt,
             controller.signal,
